+++ AppointWell/app/src/main/java/com/example/myapplication/Patient.java
@@ -1,19 +1,19 @@
 package com.example.myapplication;
 
 public class Patient extends UserAccount{
-    private int healthCardNumber;
+    private long healthCardNumber;
     public Patient(){}
 
-    public Patient(String email, String firstName, String lastName, String password, String address, int healthCardNumber, int phoneNumber) {
+    public Patient(String email, String firstName, String lastName, String password, String address, long healthCardNumber, long phoneNumber) {
         super(email, password, "Patient", "Pending", firstName, lastName, address, phoneNumber);
         this.healthCardNumber = healthCardNumber;
     }
 
-    public int getHealthCardNumber() {
+    public long getHealthCardNumber() {
         return healthCardNumber;
     }
 
-    public void setHealthCardNumber(int healthCardNumber) {
+    public void setHealthCardNumber(long healthCardNumber) {
         this.healthCardNumber = healthCardNumber;
     }
 }
